+++ test/test_library.py
@@ -240,27 +240,6 @@ class DestinationTest(_common.TestCase):
         self.assertFalse('two \\ three' in p)
         self.assertFalse('two / three' in p)
 
-    def test_sanitize_unix_replaces_leading_dot(self):
-        with _common.platform_posix():
-            p = util.sanitize_path(u'one/.two/three')
-        self.assertFalse('.' in p)
-
-    def test_sanitize_windows_replaces_trailing_dot(self):
-        with _common.platform_windows():
-            p = util.sanitize_path(u'one/two./three')
-        self.assertFalse('.' in p)
-
-    def test_sanitize_windows_replaces_illegal_chars(self):
-        with _common.platform_windows():
-            p = util.sanitize_path(u':*?"<>|')
-        self.assertFalse(':' in p)
-        self.assertFalse('*' in p)
-        self.assertFalse('?' in p)
-        self.assertFalse('"' in p)
-        self.assertFalse('<' in p)
-        self.assertFalse('>' in p)
-        self.assertFalse('|' in p)
-
     def test_path_with_format(self):
         self.lib.path_formats = [('default', '$artist/$album ($format)')]
         p = self.i.destination()
@@ -337,11 +316,6 @@ class DestinationTest(_common.TestCase):
         ]
         self.assertEqual(self.i.destination(), np('one/three'))
 
-    def test_sanitize_windows_replaces_trailing_space(self):
-        with _common.platform_windows():
-            p = util.sanitize_path(u'one/two /three')
-        self.assertFalse(' ' in p)
-
     def test_get_formatted_does_not_replace_separators(self):
         with _common.platform_posix():
             name = os.path.join('a', 'b')
@@ -407,25 +381,6 @@ class DestinationTest(_common.TestCase):
         p = self.i.destination()
         self.assertEqual(p.rsplit(os.path.sep, 1)[1], 'something')
 
-    def test_sanitize_path_works_on_empty_string(self):
-        with _common.platform_posix():
-            p = util.sanitize_path(u'')
-        self.assertEqual(p, u'')
-
-    def test_sanitize_with_custom_replace_overrides_built_in_sub(self):
-        with _common.platform_posix():
-            p = util.sanitize_path(u'a/.?/b', [
-                (re.compile(r'foo'), u'bar'),
-            ])
-        self.assertEqual(p, u'a/.?/b')
-
-    def test_sanitize_with_custom_replace_adds_replacements(self):
-        with _common.platform_posix():
-            p = util.sanitize_path(u'foo/bar', [
-                (re.compile(r'foo'), u'bar'),
-            ])
-        self.assertEqual(p, u'bar/bar')
-
     def test_unicode_normalized_nfd_on_mac(self):
         instr = unicodedata.normalize('NFC', u'caf\xe9')
         self.lib.path_formats = [('default', instr)]
@@ -473,14 +428,6 @@ class DestinationTest(_common.TestCase):
         self.i.album = 'bar'
         self.assertEqual(self.i.destination(),
                          np('base/ber/foo'))
-
-    @unittest.skip('unimplemented: #359')
-    def test_sanitize_empty_component(self):
-        with _common.platform_posix():
-            p = util.sanitize_path(u'foo//bar', [
-                (re.compile(r'^$'), u'_'),
-            ])
-        self.assertEqual(p, u'foo/_/bar')
 
     @unittest.skip('unimplemented: #359')
     def test_destination_with_empty_component(self):
@@ -698,49 +645,6 @@ class DisambiguationTest(_common.TestCase, PathFormattingMixin):
         album1.store()
         self._setf(u'foo%aunique{albumartist album,albumtype}/$title')
         self._assert_dest('/base/foo [foo_bar]/the title', self.i1)
-
-
-class PathConversionTest(_common.TestCase):
-    def test_syspath_windows_format(self):
-        with _common.platform_windows():
-            path = os.path.join('a', 'b', 'c')
-            outpath = util.syspath(path)
-        self.assertTrue(isinstance(outpath, unicode))
-        self.assertTrue(outpath.startswith(u'\\\\?\\'))
-
-    def test_syspath_windows_format_unc_path(self):
-        # The \\?\ prefix on Windows behaves differently with UNC
-        # (network share) paths.
-        path = '\\\\server\\share\\file.mp3'
-        with _common.platform_windows():
-            outpath = util.syspath(path)
-        self.assertTrue(isinstance(outpath, unicode))
-        self.assertEqual(outpath, u'\\\\?\\UNC\\server\\share\\file.mp3')
-
-    def test_syspath_posix_unchanged(self):
-        with _common.platform_posix():
-            path = os.path.join('a', 'b', 'c')
-            outpath = util.syspath(path)
-        self.assertEqual(path, outpath)
-
-    def _windows_bytestring_path(self, path):
-        old_gfse = sys.getfilesystemencoding
-        sys.getfilesystemencoding = lambda: 'mbcs'
-        try:
-            with _common.platform_windows():
-                return util.bytestring_path(path)
-        finally:
-            sys.getfilesystemencoding = old_gfse
-
-    def test_bytestring_path_windows_encodes_utf8(self):
-        path = u'caf\xe9'
-        outpath = self._windows_bytestring_path(path)
-        self.assertEqual(path, outpath.decode('utf8'))
-
-    def test_bytesting_path_windows_removes_magic_prefix(self):
-        path = u'\\\\?\\C:\\caf\xe9'
-        outpath = self._windows_bytestring_path(path)
-        self.assertEqual(outpath, u'C:\\caf\xe9'.encode('utf8'))
 
 
 class PluginDestinationTest(_common.TestCase):
@@ -1002,23 +906,6 @@ class PathStringTest(_common.TestCase):
         )
         alb = self.lib.get_album(alb.id)
         self.assert_(isinstance(alb.artpath, bytes))
-
-
-class PathTruncationTest(_common.TestCase):
-    def test_truncate_bytestring(self):
-        with _common.platform_posix():
-            p = util.truncate_path(b'abcde/fgh', 4)
-        self.assertEqual(p, b'abcd/fgh')
-
-    def test_truncate_unicode(self):
-        with _common.platform_posix():
-            p = util.truncate_path(u'abcde/fgh', 4)
-        self.assertEqual(p, u'abcd/fgh')
-
-    def test_truncate_preserves_extension(self):
-        with _common.platform_posix():
-            p = util.truncate_path(u'abcde/fgh.ext', 5)
-        self.assertEqual(p, u'abcde/f.ext')
 
 
 class MtimeTest(_common.TestCase):
